Fix signedness errors in format strings.

+++ src/smpi/mpi/smpi_f2c.cpp
@@ -32,12 +32,12 @@ int F2C::f2c_id(){
 };
 
 char* F2C::get_key(char* key, int id) {
-  std::snprintf(key, KEY_SIZE, "%x",id);
+  std::snprintf(key, KEY_SIZE, "%x", (unsigned)id);
   return key;
 }
 
 char* F2C::get_key_id(char* key, int id) {
-  std::snprintf(key, KEY_SIZE, "%x_%d",id, smpi_process()->index());
+  std::snprintf(key, KEY_SIZE, "%x_%d", (unsigned)id, smpi_process()->index());
   return key;
 }
